Asking order, can't divide by 0
I changed that if the option you choose is not valid (not from 1 to 4) then it says it's not valid isntead of asking for two numbers and then saying it's not valid.

I changed that when you choose divide, you can't divide by 0.

I also added the versions list to the README